third_party/WebKit/LayoutTests/media/mediasession/mojo/media-control-action-reaches-client.html

<!DOCTYPE html>
<title>MediaSession Mojo Test</title>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/mojo-helpers.js"></script>
<script src="resources/mediasessionservice-mock.js"></script>
<script src="resources/utils.js"></script>
<script>

var mock;

var expectations = [
  "play",
  "pause",
  "playpause",
  "previoustrack",
  "nexttrack",
  "seekbackward",
  "seekforward",
];

var nextExpectation = 0;

function checkExpectation(t, event) {
  var expectedEvent = expectations[nextExpectation];
  assert_equals(expectedEvent, event.type);
  if (++nextExpectation >= expectations.length)
    t.done();
}

function runTests(t) {
  window.navigator.mediaSession.onplay = t.step_func(checkExpectation.bind(null, t));
  window.navigator.mediaSession.onpause = t.step_func(checkExpectation.bind(null, t));
  window.navigator.mediaSession.onplaypause = t.step_func(checkExpectation.bind(null, t));
  window.navigator.mediaSession.onprevioustrack = t.step_func(checkExpectation.bind(null, t));
  window.navigator.mediaSession.onnexttrack = t.step_func(checkExpectation.bind(null, t));
  window.navigator.mediaSession.onseekforward = t.step_func(checkExpectation.bind(null, t));
  window.navigator.mediaSession.onseekbackward = t.step_func(checkExpectation.bind(null, t));

  mock.getClient().didReceiveAction(MediaSessionAction.PLAY);
  mock.getClient().didReceiveAction(MediaSessionAction.PAUSE);
  mock.getClient().didReceiveAction(MediaSessionAction.PLAY_PAUSE);
  mock.getClient().didReceiveAction(MediaSessionAction.PREVIOUS_TRACK);
  mock.getClient().didReceiveAction(MediaSessionAction.NEXT_TRACK);
  mock.getClient().didReceiveAction(MediaSessionAction.SEEK_BACKWARD);
  mock.getClient().didReceiveAction(MediaSessionAction.SEEK_FORWARD);
}

// Use async_test to do asynchronous setup since setup() only works for
// synchronous setup.
async_test(function(t) {
  mediaSessionServiceMock.then(m => {
    mock = m;
    mock.setClientCallback(t.step_func(runTests.bind(null, t)));
    // Touch window.navigator.mediaSession to start the service.
    window.navigator.mediaSession.metadata = null;
  });
}, "test that the mock service is setup");

</script>
